feat: Rebrand Rubberduck to ThetaEdgeX for release readiness
The extension is rebranded from 'rubberduck-vscode' to 'thetaedgex-vscode' across various files, including the root package.json, .all-contributorsrc and in-app descriptions and prompts. Extensive alterations have been made to app/vscode/asset/package.json and app/vscode/asset/walkthrough/setup.md to better reflect the new brand and change configuration setups. This refresh is in anticipation of a wider release, and to ensure brand identity consistency across the application. Note that this does not alter any core functionalities or architecture.

Closes #519.